+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# 0.2.1 (30. September, 2024)
+
+Expose a signal that can be awaited on without awaiting the configured
+delay first. If no delay is used this API is equivalent to the already
+existing `cancelled` function.
+
+This can be used for scenarios where you do not need a graceful buffer and would like to
+cancel as soon as a signal is received.
+
 # 0.2.0 (29. September, 2024)
 
 This is usability wise not a breaking release,